Clean S9API use
